refactor: sVG 파일 경로 변경 및 새로운 아이콘 컴포넌트 추가
