+++ spark-web/src/assets/svg/logo/Analysis.tsx
@@ -1,0 +1,29 @@
+import { SvgProps } from '../nav/SettingIcon';
+
+export const Analysis = ({ ...props }: SvgProps) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="25"
+      height="24"
+      viewBox="0 0 25 24"
+      fill="none"
+      {...props}
+    >
+      <path
+        d="M15.6883 15.2943C14.6069 16.3497 13.1284 17 11.498 17C9.86374 17 8.38207 16.3466 7.29999 15.2868C7.73893 14.3307 8.70482 13.6667 9.82578 13.6667H13.1591C14.283 13.6667 15.251 14.3341 15.6883 15.2943Z"
+        fill="white"
+      />
+      <path
+        d="M11.4924 12.5556C13.0266 12.5556 14.2702 11.3119 14.2702 9.77778C14.2702 8.24365 13.0266 7 11.4924 7C9.95832 7 8.71467 8.24365 8.71467 9.77778C8.71467 11.3119 9.95832 12.5556 11.4924 12.5556Z"
+        fill="white"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M16.4056 17.3199C15.0509 18.3729 13.3487 19 11.5 19C7.08172 19 3.5 15.4183 3.5 11C3.5 6.58172 7.08172 3 11.5 3C15.9183 3 19.5 6.58172 19.5 11C19.5 12.8488 18.8729 14.551 17.8198 15.9057L21.2071 19.293C21.5976 19.6836 21.5976 20.3167 21.2071 20.7073C20.8166 21.0978 20.1834 21.0978 19.7929 20.7073L16.4056 17.3199ZM17.5 11C17.5 14.3137 14.8137 17 11.5 17C8.18629 17 5.5 14.3137 5.5 11C5.5 7.68629 8.18629 5 11.5 5C14.8137 5 17.5 7.68629 17.5 11Z"
+        fill="white"
+      />
+    </svg>
+  );
+};
