+++ spark-web/src/assets/svg/logo/Bulb.tsx
@@ -1,0 +1,21 @@
+import { SvgProps } from '../nav/SettingIcon';
+
+export const Bulb = ({ ...props }: SvgProps) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="25"
+      height="24"
+      viewBox="0 0 25 24"
+      fill="none"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M7.55025 4.05025C8.86301 2.7375 10.6435 2 12.5 2C14.3565 2 16.137 2.7375 17.4497 4.05025C18.7625 5.36301 19.5 7.14348 19.5 9C19.5 11.38 18.31 13.47 16.5 14.74V17C16.5 17.2652 16.3946 17.5196 16.2071 17.7071C16.0196 17.8946 15.7652 18 15.5 18H9.5C9.23478 18 8.98043 17.8946 8.79289 17.7071C8.60536 17.5196 8.5 17.2652 8.5 17V14.74C6.69 13.47 5.5 11.38 5.5 9C5.5 7.14348 6.2375 5.36301 7.55025 4.05025ZM9.79289 21.7071C9.60536 21.5196 9.5 21.2652 9.5 21V20H15.5V21C15.5 21.2652 15.3946 21.5196 15.2071 21.7071C15.0196 21.8946 14.7652 22 14.5 22H10.5C10.2348 22 9.98043 21.8946 9.79289 21.7071ZM14.528 9.84722C14.9829 9.67851 14.9845 9.03567 14.5304 8.86468L13.5018 8.47728C13.4552 8.45973 13.4128 8.43244 13.3776 8.39723C13.3396 8.35919 13.3108 8.31289 13.2936 8.26194L12.9432 7.22669C12.7836 6.75494 12.1181 6.75005 11.9515 7.2194L11.5473 8.35842C11.5301 8.40694 11.5023 8.451 11.4659 8.4874C11.429 8.52424 11.3844 8.55228 11.3352 8.56945L10.4816 8.86734C10.0137 9.03061 10.0124 9.69176 10.4795 9.85697L11.4248 10.1912C11.4638 10.205 11.4993 10.2274 11.5285 10.2567C11.5602 10.2884 11.5838 10.3273 11.5972 10.37L11.9514 11.5022C12.1054 11.9945 12.8028 11.9921 12.9535 11.4989L13.2808 10.4274C13.2942 10.3835 13.3182 10.3436 13.3507 10.3111C13.3792 10.2826 13.4135 10.2606 13.4514 10.2465L14.528 9.84722Z"
+        fill="white"
+      />
+    </svg>
+  );
+};
